got data of question 6.

+++ src/util/adaptor/FpGrowthProjectAdaptor.java
@@ -3,6 +3,7 @@ package util.adaptor;
 import database.DatabaseOperator;
 import fpalgorithm.fsetgenerator.fpgrowth.Algorithm;
 import fpalgorithm.fsetgenerator.fpgrowth.util.FpListItem;
+import fpalgorithm.fsetgenerator.util.result.FrequentSet;
 import fpalgorithm.fsetgenerator.util.result.FrequentSetContainer;
 import fpalgorithm.fsetgenerator.fpgrowth.util.Transaction;
 import javafx.util.Pair;
@@ -11,6 +12,9 @@ import util.GlobalInfo;
 import java.util.*;
 
 public class FpGrowthProjectAdaptor {
+//    public static FrequentSetContainer runFpGrowth() {
+//        return runFpGrowth(GlobalInfo.total_support);
+//    }
     public static FrequentSetContainer runFpGrowth() {
         Pair<ArrayList<FpListItem>, HashMap<Integer, Integer>> database_info = DatabaseOperator.getFpItemCount();
         List<FpListItem> orderedItemSet = database_info.getKey();
@@ -18,7 +22,7 @@ public class FpGrowthProjectAdaptor {
 
         List<Transaction> transactions = DatabaseOperator.getTransactions();
 
-        return Algorithm.calculateFrequentSets(orderedItemSet, transactions, GlobalInfo.total_support,
+        return Algorithm.calculateFrequentSets(orderedItemSet, transactions, (int) GlobalInfo.total_support,
                 item_orderMap);
     }
 }
